refactor: header.js 파일 분리: header과 tab, dateInfo를 분리
