+++ components/calendar/calendar.js
@@ -1,9 +1,9 @@
 import {
   getTotalIncomeData,
   getTotalExpenseData,
-} from "../utils/transaction.js";
-import { formatMoney } from "../utils/format.js";
-import { dateStore, transactionStore } from "../store/index.js";
+} from "../../utils/transaction.js";
+import { formatMoney } from "../../utils/format.js";
+import { dateStore, transactionStore } from "../../store/index.js";
 
 const WEEKDAYS = ["일", "월", "화", "수", "목", "금", "토"];
