+++ components/main/transactionsList.js
@@ -1,11 +1,11 @@
 import {
   getTotalIncomeData,
   getTotalExpenseData,
-} from "../utils/transaction.js";
-import { CATEGORY_NAME } from "../constants/category.js";
-import { formatMoney } from "../utils/format.js";
+} from "../../utils/transaction.js";
+import { CATEGORY_NAME } from "../../constants/category.js";
+import { formatMoney } from "../../utils/format.js";
 import { fillFormWithTransaction, cancelEditMode } from "./inputBar.js";
-import { dateStore, transactionStore } from "../store/index.js";
+import { dateStore, transactionStore } from "../../store/index.js";
 
 // 클릭된 행 상태 관리
 let selectedRowId = null;
